+++ sites/pc/entry/index.tsx
@@ -1,5 +1,4 @@
 import React, { createContext, useEffect } from 'react';
-import ReactDOM from 'react-dom';
 import { HashRouter, Switch, Route } from 'react-router-dom';
 import demoDocs from '../pages/components';
 import enDemoDocs from '../pages/components/index-en-US';
@@ -7,6 +6,7 @@ import Demo from './demo';
 import Home from './home';
 import readmeDocs from '../pages/guide';
 import { LanguageSupport } from '../../utils/language';
+import { render } from '../../../packages/arcodesign/components/_helpers';
 import { showGA } from '../../utils/ga';
 import './index.less';
 
@@ -109,4 +109,4 @@ function App() {
         </div>
     );
 }
-ReactDOM.render(<App />, document.querySelector('#root'));
+render(<App />, document.querySelector('#root')!);